agri-cows-api/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
marshmallow==3.21.3
marshmallow-sqlalchemy==1.0.0
Werkzeug==3.0.3
Flask-Cors==4.0.1
